Refactor dashboard aggregation into cached shared server utility
