+++ src/app/api/accessories/[id]/battery/route.ts
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { prisma } from "@/lib/prisma";
+import { revalidateDashboardCaches } from "@/lib/server/dashboard";
 
 // POST /api/accessories/[id]/battery - Log a battery change
 export async function POST(_req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
@@ -13,6 +14,8 @@ export async function POST(_req: NextRequest, { params }: { params: Promise<{ id
     where: { id },
     data: { batteryChangedAt: new Date() },
   });
+
+  revalidateDashboardCaches(["accessories"]);
 
   return NextResponse.json(updated);
 }
