+++ src/app/api/ammo/[id]/transactions/route.ts
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { prisma } from "@/lib/prisma";
+import { revalidateDashboardCaches } from "@/lib/server/dashboard";
 
 // Types that subtract from quantity
 const SUBTRACT_TYPES = new Set(["RANGE_USE", "TRANSFER_OUT", "EXPENDED"]);
@@ -100,6 +101,8 @@ export async function POST(
       }),
     ]);
 
+    revalidateDashboardCaches(["ammo", "range"]);
+
     return NextResponse.json(
       {
         stock: updatedStock,
@@ -109,6 +112,7 @@ export async function POST(
     );
   } catch (error) {
     console.error("POST /api/ammo/[id]/transactions error:", error);
+
     return NextResponse.json(
       { error: "Failed to create ammo transaction" },
       { status: 500 }
