+++ src/app/api/firearms/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { prisma } from "@/lib/prisma";
 import { encryptField, decryptField } from "@/lib/crypto";
+import { revalidateDashboardCaches } from "@/lib/server/dashboard";
 
 // GET /api/firearms - List all firearms with build count
 export async function GET() {
@@ -97,6 +98,8 @@ export async function POST(request: NextRequest) {
         },
       },
     });
+
+    revalidateDashboardCaches(["firearms"]);
 
     return NextResponse.json(
       { ...firearm, buildCount: firearm._count.builds, _count: undefined },
